定义App接口，接口暂未实现业务逻辑

+++ shop-user/shop-user-api/src/main/java/quick/pager/shop/user/model/UserInfo.java
@@ -7,6 +7,8 @@ import lombok.EqualsAndHashCode;
 import quick.pager.shop.model.Model;
 
 /**
+ * 用户基本信息
+ *
  * @author siguiyang
  */
 @EqualsAndHashCode(callSuper = true)
@@ -15,18 +17,35 @@ public class UserInfo extends Model {
 
     private static final long serialVersionUID = 788660856242390694L;
 
+    /**
+     * 用户主键
+     */
     private Long userId;
-
+    /**
+     * 昵称
+     */
     private String username;
-
+    /**
+     * 性别
+     * true 男，false 女
+     */
     private Boolean gender;
 
-    private Byte age;
-
+    /**
+     * 年龄
+     */
+    private Integer age;
+    /**
+     * 邮箱地址
+     */
     private String email;
-
+    /**
+     * 生日
+     */
     private Date birthday;
-
+    /**
+     * 头像地址
+     */
     private String avatar;
 
 }
